+++ src/api/app_routes/__init__.py
@@ -1,0 +1,1 @@
+from .user_authenticated import apiAuthUser
